requirements.txt
inquirer==3.1.1
Pillow==9.4.0
python-decouple==3.6
PyYAML==6.0
requests==2.28.1
selenium==4.7.2
setuptools==62.6.0
yt_dlp==2021.6.1
